WEB-4385: Switching from SHA256 to MD5 hashing algorithm
SHA256 is 256 bits = 64 (hex) characters
MD5 is 128 bits = 32 (hex) characters

So the hash in the filename should be half the length, which I think
is ok.

Note: every single video course and book that runs through CI after
this will re-upload all its images. I don't think that this is a
problem, but something to bear in mind. The old images will remain
on the CDN, so it shouldn't break anything. But we'll have duplicates
with different names.