WEB-4523: Fixes a problem with clashing method names
Basically, some files we check both attached images and markdown images
which means that we use both the Markdown and the Attachable modules.
Trouble is, they shared some method naming, which means that we
overwrote one with the other.

This gives the similar methods in the two modules unique names, so they
no longer clash.

Also updates the error message to be more helpful.